Support omdbapi.com authentication
omdbapi.com has transitioned to a private model with mandatory
authentication. Each person wanting to use the account has to
create a pledge to omdbapi.com in order to acquire the apikey.

Once a user has an apikey the apikey is set in the application
and added to the omdbapi.com queries as a query parameter.

Added a new omdb main module which creates an omdb setting widget.

Added support for apikey in query movie database query parameters
and a new test function to test the apikey.

Issue: #57