Upgraded to OpenSSL 1.0.2k
Upgraded to OpenSSL 1.0.2k.
Started using OpenSSL from third_party/openssl also for the Linux
build since this makes the environment more homogenous. Morever
ArchLinux already moved onto OpenSSL 1.1.x which is not compatible
with OpenSSL 1.0.x.

Note that this means that now Qt4 needs to be build with openssl from
third_party also on Linux not only on Windows.