Improve the header file data usage
The header file data blocks are wasting space for groups with high
volume of multipart binaries.

We can improve on this by changing the way we store items in the
catalog files. Instead of indexing into the catalogs by hash modulo
we can fill every bucket in the catalog by generating an internal
article number and packing them items densely into the catalogs.

When first header is downloaded we grab the article number as a landmark.
The landmark divides the range of headers into "newer" (referred to as
positive) and "older" (referred to as negative) ranges. Then using two
counters that give the number of items in both direction we can fill
every slot by generating a unique internal article number by comparing
the article's number to the landmark and then adding either the positive
or negative offset and then incrementing the offset.

Additionally we don't need to keep track of hash codes for single part
binaries. We only need the hash code tracking for multipart binaries
and the hash code should now map to an internal article number. When a
hash match happens we can then easily navigate to the actual article
by the internal article number.

Issue #59