Update zlib version and building steps
Update zlib version and building steps. Zlib is now at version 1.2.11.

Also in anticipation of moving to a non boost.build based build the
zlib build is now done with the CMake that comes with it.

Additionally removed a double zlib from the build process. The engine
was using zlib through a library called libzlib.so and Qt was using
the system installed libz. These were both different versions of zlib
and it's never a good idea to depend on 2 different versions of the same
library and have these loaded in the process.

On Linux Qt must now be built after the zlib build and it needs to
use/link the zlib found in third_party/zlib.