First version of a simple SMTP client implementation.
First version of a simple SMTP client implementation. This uses the
LGPL library by Tokes Attila.
https://github.com/bluetiger9/SmtpClient-for-Qt

The SMTP client is unfortunately using blocking Qt socket methods
such as waitForReadyRead which blocks the UI and makes it hard
to cancel the operation. This needs more refactoring, but the
core functionality is there, i.e. it's now possible to setup a
SMTP email configuration and send email.