Add a background SmtpClient thread to avoid blocking the application
Add a background SmtpClient thread to avoid blocking the application
and the the UI.

The actual problem is that the third_party/smtpclient uses the blocking
Qt socket API for waiting socket readability and writeability. Calling
these methods from the applications main thread will obviously block
the UI and the rest of the application logic from running.

We can mitigate the issue by offloading the SMTP acitivity to a thread
that runs and blocks in the background. This keeps rest of the application
responsive.

A possibly better solution would be to refactore the smtpclient to
use non-blocking socket API but this is left as a future exercise.