Added a factory interface to create engine's tasks and connections
Added a factory interface to create engines' tasks and connections.
This is first step in refactoring the engine (and the relevant state
classes) to be completely implementation independent state machine.

Further work is required to refactor to engine to be completely
independent from any specific task and connection implementations.
Once this is done the engine's state machine logic can easily be tested.