Always commit the update task status to files
Always commit the update task status to files, otherwise the state
will be inconsistent and corrupted.

We're now doing this commit from the dtor to make sure that it's
always called but this has the problem of suppressing any errors
that might occur. The engine should be refactored to perform
a proper shutdown for tasks so that errors could be propagated to the
UI layer.

Issue #79