Move update store actions to a separate thread instead of UI thread
Move update store actions to a separate thread in the threadpool
instead of performing these on the UI thread. The problem is that the
store operation (which updates the database files and writes changes)
is a rather slow operation and doing this in on the UI will cause
bad janks.

However the UI needs to be able to have a consistent view of the data
so that it can open the catalog/data files safely. We can achieve this
by having an ugly mutex that that the UI thread implicitly acquires in
the update when it pumps the engine's queues and dispatches the actions.
We then invoke the callbacks only after acquiring the mutex, which makes
sure that the background thread cannot writing catalog header data while
the UI thread is reading it thus resulting in inconsistent state.

This design is rather error prone and ugly. A possibly alternative way
to implement this would be to have the datafiles create a "snapshot"
state object that is dispatched to the UI which then uses the snapshot
state to open the datafiles. This would not only be a safe design but
would also allow the background operation to proceed in parallel.