Refactored some TaskId related functionality
Refactored some TaskId related functionality. The TaskId is a bit
messy currently since the engine gives back actually the batch id
and the the KillTaskById works on the batch level as well. Yet there
exists an API such as GetActionId (incorrectly named) which gives
a task id for task instead of batch if item grouping is not set.

This refactoring exposes TaskId only for batch items. The complicated
GetActionId API is then removed and functions such as Kill/Pause/Resume
which take a list index return a TaskId but only IFF item grouping is set.

This is still a bit too complicated but slightly better now than before.