Open catalog from the snapshot in the UI whenever possible
Open catalog from the snapshot in the UI whenever possible. When the
update process is running the engine can provide the UI with a list
of snapshots which can be used to open the UI's catalog objects
in order to have a consistent view and without having to lock the task.