Refactored connection class to hide implementation details
Refactored connection class to hide implementation details such as
the action objects that it performs when connecting, initialiazing
NNTP session etc. These are really internal to the connection
and if possible should not be in the public interface.

Added new connection functions to retrieve the current connection
state and error (if any).

Refactored the engine to not depend on the connection class's
action details but only on the public interface.

Refactored the cmdlist completion into a callback which is registered
with the connection and that gets then invoked through an action
when run_completion_callbacks is called.