Refactored the engine log creation into the factory interface
Refactored the engine log creation into the factory interface. This
is convenient for testing purposes.

This means that for the "default" case (when the engine uses a real
internal factory implementation) the log path only makes sense
when constructing the engine. Because of this the engine creation
in the application is moved into load state when we have the log
path available.