Refactored the task exception handling a little bit
Refactored the task exception handling. Currently it's not
excepted that we can handle cmdlist exceptions meaningfully,
so we'll let those propagate.

However the exceptions that are interested are the exceptions that
migth have occurred during some background action. For example a file
write has failed etc. We handle these exceptions and transition the
task into error state.