Added general NNTP response handling in the Session class
Added general NNTP response handling in the Session class.
There are some general responses that can occur with any command
and we can handle them in a generic fashion instead of each command
having to understand them.

The responses are in 1xx, 2xx, 4xx and 5xx category. Currently
we're only interested in 502 response which is used to indicate
"no permission" (access restriction) kind of situation. If this
general response is sent then the session goes into NoPermission
error state.