Refactored engine task list saving
Refactored engine task list saving. Simplified the implementation
in the sense that it *only* saves the list of tasks. Other state
such as accounts or other settings not saved.

Refactored Task's Pack and Load methods so that they operate on
opaque byte streams only. The byte stream is then managed by the
engine when the task list is saved and loaded.

Added engine unit test case to load and save tasklist.