Update task progress in action completion callback
Update task progress in action completion callback. Previously there
was a bug that updated the progress in the method to update the incoming
state object leaving the internal state as zero progress. This would
mean that when the completion callback fired the progress/completion
was also zero instead of the actual value.