Created a windows specific WebWidget that can load advertisements
Created a windows specific WebWidget that can load web content from
a URL. This is based on a Windows specific ActiveX (COM) technology.

The problem is that the Usenext advertisements currently are provided
as obfuscated JS link which loads a JS script from their website and
then creates the appropriate web page content at runtime. So basically
in order to correctly link to the advertisement campaigns the JS
must be evaluated. So the current solution is to use a component that
can load the HTML page from the web and evaluate the JS on that page
and then finally render the advertisement appropriately.

The downside of the current implementation is that this is specific
to Windows and doesn't work portably. Seems that a similar
implementation for Linux would require building Qt with WebKit.
Currently we're disabling all Linux adverts.