Optimized the engine's filetype identification method
Optimized the engine's filetype identification method. This is a
simple method that expects to be given a filename with the extension
string and then maps the extension to a file type category.

After profiling unit_test_update on Valgrind (callgrind) the indication
was that considerable time was spent in find_filetype method constructing
and running the regulex expressions to identify the filetype category.
However looking at the code it seems that we can simply optimize away the
regular expression by pre-constructing a simple extension -> type map and
expand the few expressions such as .r\\d{1,3} directly into strings such
as .r00, .r01, .r02 etc.

After doing this change the find_filename method completely disappears
from the profiling information's  most expensive methods.