Optimized the article parsing by skipping is_binary_post call
Optimized the article parsing by skipping is_binary_post call.
After spending some time profiling the unit_test_update there was
a clear indication that nntp:is_binary_post and nntp::find_filename
are both quite expensive calls and considerable time is spent in
these two functions.

However we can simplify the logic by skipping the call to is_binary_post
and by directly looking for a find_filename. If a filename is found
we can then assume that the post is indeed a binary post and proceed
as before.

Overall this change eliminates one of these expensive calls and results
(with the changes done to the find_filetype) in about 80% runtime
reduction in unit_test_update testcase.