Renamed CmdList::InspectRawBuffer to InspectIntermediateContentBuffer
Renamed CmdList::InspectRawBuffer to InspectIntermediateContentBuffer.
This is more appropriate since we're really interested in having a look
at the intermediate *content* not the actual raw socket data.

Also added a check in the connection to only call this method when
the currently executing session command is a command that is currently
transferring content.