Support intermediate LIST buffer parsing and content display on UI
Support intermediate LIST buffer parsing and displaying the intermediate
newsgroup listing contents on the UI.

The problem is that the LIST operation is often very slow and it takes
several minutes to obtain the full response body to LIST command.

The CmdList now supports a mechanism to inspect the intermediate content
buffer as it's getting downloaded. It's possible to use this mechanism
in the Listing task to parse the content buffer while it's still running
and start parsing the group list data as soon as possible and provide
this to the application layer which can then display this intermediate
data without having to wait for the whole operation to complete.