Refactored Newznab implementation to also support RSS feeds
Refactored Newznab implementation to also support RSS feeds.

Newznab sites provide RSS feeds. We currently had a "separate"
RSS account system and sort of implementation specific to nzbs.org,
which really is just the same as implementing it generally for
*any* Newznab site. Therefore the nzbs.org RSS can be refactored
into a generic RSS implementation targeting Newznab servers.

This is a large commit and implements the following changes:
- Removed womble implementation. Site seems to have gone down.
- Refactored "Search" module (in GUI) into Newznab
- Used the nzbs.org RSS implementation as a baseline for generic
  Newznab RSS implementation.
- Removed specific RSS Settings (Part of Newznab settings now)
- Multiple RSS windows can be now opened through the MainWindow
  (similar to search)
- Added UserId to Newznab account. Needed by the RSS queries.