Support for optionally only adding servers in Newznab import
Support for optionally only adding servers in Newznab import.
Importing the list of newznab servers used to try to do a registration
by default. But it's possible that user just wants to add them for
now and perhaps register manually (or has already registered).

Additionally only the servers that are selected in the list are either
added or tried to register an account with.