Added timeout detection in the Process class
Added timeout detection in the Process class. The assumption is that
the 3rd party applications (such as unrar, par2) should produce some
output either in the stdout or in the stderr IO channel. If this there's
no input within a time interval (30s currently) we assume that the
process is in fact hung/frozen.

Also Simplified the Process::kill implementation slightly so that
when kill is called the onFinished signal no longer runs. This made
the logic in unrar/par2/unzip also simpler.

Added the appropriate error message for timeout in unrar/par2/unzip

Issue #96