Fixed assertion when exiting application with active unrar
Fixed assertion when exiting application with active unrar.
The underlying cause is that shutdown process would simply call
stopUnpack which would invoke the finished callback which would
would (inadvertedly) schedule another Archive for extraction.
Thus there'd be an ective unrar process and Process class would
assert in the dtor.

This patch fixes the bug by organizing the code so that there's
a new shutdown method that gets called on application shutdown.
This will simply clean up all the processing and not run any signals
or schedule any new extracts.

Similar bug existed with also the unzip and par2 implementations.

Issue #98