New task interface method to check for reliable progress information
Sometimes the task (for example the Update task) cannot compute
progress information properly until at some later point for
example after some cmdlist has run.

Subsequently the BatchState in Engine was incorrectly computing
partial completions based on this bogus data.

Added a new method in the Task interface to check if the progress
information is reliably available.

Added a new flag in the ui/TaskDesc to indicate to the UI whether
the completion percentage is reliable or not.

Issue #97