Refactor session widget loading and saving
Refactor session widget loading and saving. These are widgets that
are opened at runtime through File|New Search or File|New RSS.
These widgets are now all saving their state to their own settings
file if they're open when the application is exited. If they're
closed while the application is running their state is not stored
anywhere and they're (obviously) not restored on the next application
run.

An exception to the above rule is the Newsgroup widget since the state
of browsing a newsgroup can be persisted per newsgroup so it's possible
to continue where the last browsing session ended even if the tab is
closed.

Issue #99