Show the number of new messages in the newsgroup list
Show the number of new messages in the newsgroup list since the
latest refresh of the newsgroup(s) with contents.

This feature is supported by having the newslist listen to group
update/update complete events coming from the engine and then writing
the current last local article number to a text file. When the
newsgroup list is then loaded for each group that has data we load
the data from the text file possible and compare the last local
article number to the last remote article number. The difference
is the number of new articles added to the group since any previous
group update/refresh.

Issue #102