Fix potentially incorrect newsgroup progress computation
Fix potentially incorrect newsgroup progress computation when the
local database has items older than what the reserver currently has
the server might report a first article number that is *greater*
than what we have already locally. This meanss that the number of
messages locally can actually be greater than what the server has
remotely. This will throw off the progress computation in the
engine (task progress) and also in the newsgroup UI.

The simple fix here is to clip the first article number to greater
of the two nubmbers, local/remote and use that.

Issue #103