Refactor drag and drop handling and session widget creation
Refactor drag and drop handling and session widget creation.
Instead of having the MainModule potentially create a new MainWidget
we can just ask the MainModule(s) to handle the drag and drop
event. If any module instances handle the event we're done. If not
then we can ask all the permanent widgets to handle the event.
If any do then we're done. Lastly we can check if the drag and drop
file types are something we know off explicitly in the MainWindow
(such as .nzb files) and then create a new session widget object.

commit removes the last dependency to MainWidget in MainModule.