Add a work around for obfuscated downloads without single filename
Add a work around for obfuscated downloads without single filename.
Some content is obfuscated so that the filenames in the yEnc headers
are randomly generated junk. This means that the code which tries
to map the decoded chunks into files on the disk creates a new file
for each chunk instead of writing the chunk into a single file in the
right offset.

This obfuscation method seems be applicable only to multipart yEnc
encoded content.

The work around here is to ignore the filename that is present in the
yEnc part headers and instead always rely on a filename coming from
some external source such as the subject line. If the subject line
contains a recognized filename we turn on a bit that tells the download
task to ignore the filenames in the yEnc headers.