Replace protobuf based serialization with nlohmann json serialization
Replace protobuf based serialization with nlohmann json serialization.

This is the glorious day when we praise God, Jesus the Merciful and
Allah, Vishnu etc. and bask in the glory of God and rejoice that we
can for once and for all rid ourselves of this horrible steaming
pile of puke that is the Google protobuf. Begone you horrible
hellish depedency and blessed be a single header file json library!