Hide OpenSSL symbols inside a binary firewall library
Hide OpenSSL symbols inside a binary firewall library. The engine
uses OpenSSL 1.0.x while for example Qt5 uses OpenSSL 1.1.x. These
are not compatible and if the OpenSSL symbols are publicly accessible
in the engine library there'll be problems at runtime.

The current solution is to create a wrapper library for all the socket
code (incl. both TCP and TCP/SSL) and link that library against SSL
libraries while instructing the linker to hide all symbols *except*
for one symbol which is used to create a new newsflash::Socket object.

This hack solution allows multiple versions of OpenSSL to exist in
the same process. The long term solution is to get rid of the fucking
OpenSSL completely anyway since the whole library is *brain damaged*.