Add a workaround for newslist sorting not working via sortByColumn
Add a workaround for newslist sorting not working via sortByColumn.
For some obscure reason in this Qt5 the request to resort the model
through QTableView::sortByColumn no longer actually invokes the
model's sort function. It looks like there's some caching where
the QTableView skips the call to model's sort because it thinks
the model is already sorted by that criteria. The current workaround
is simply to call the model's sort directly.