Fix bugs when filling articles after not being able to select group
Fix bugs when filling articles after not being able to select group.
The manifestation of the bug is that when the main server doesn't
have the newsgroup in question filling doesn't work properly. There
are actually relevant 2 issues here.

1st issue is that the session incorrectly caches the current group
event though the request to change group has failed. This means that
on the next cmd list the group change command is skipped since it's
incorrectly assumed to be a redundant state change.

2nd issue is that the failure to select the newsgroup turns on the
failbit in the cmd list which isn't cleared. So when the cmd list
runs against the fill server the fail bit is still on regardless of
whether the fills complete or not. The download task then skips all
processing of the buffers on fail bit.

Fix failure to fill articles from the